Sorting images to place them in collage
Introduction
Images' computational processing and manipulation become easier to understand by bit mapping their representation, in other words, looking at images as “grids or rectangles of pixels” (Montfort, 2016, p. 159).
Problem definition: Approach to solving the given problem and tasks
The Problem starts with loading all the images from the folder and then processing them and get the average value of the R channel to accordingly sort them in a list according to the average value of the R channel.
after getting a sorted list that contains both images as objects and the average value of their R channel, Images need to be put in a 3 x 3 grid, in other words, an image collage according to the aforementioned ascending order.
Methodology approach: 
I started with defining the function that calculates the average of R channel. It includes getting the size of the image using the getpixel function (ibid, p.4). I then defined a function that opens all images in the folder using the globe library (ibid, p. 192). the images are loaded to two lists according to their image type (JPG and PNG) then to one list as objects. I looped through each image to get its average value using the first function within this function. The average R values were appended to another list and both lists to a nested list of images and their values sub-lists.
The ‘create a collage’ function started with getting the size of images (ibid, p. 168). Then a scaling factor in which the numbers of images by row were defined according to the frame width. According to the defined scaling factor, images width and height were scaled using the ceil and floor functions from the math library, in order to be able to place them in the frame properly by taking the number of pictures per row and the number of pixels for the width of the collage and calculating accordingly the frame height.
after the calculations, I defined the chords to place thumbnails (resized images) to them according to the previous calculations using a for loop with an if statement to iterate through a 3 x 3 grid and save the collage in a PNG format.
Outcomes/Analysis of results
I started by calling the image list function I previously defined to open all of the images in the directory folder and get their R channel average value and accordingly get a nested list. I then use the sort function to sort the image list and used a Lambda function that will use the average value of the art channel as a key to sort the list items.
The sorted list is still a nested list with both images and average values as sublists. to get the images in a flat list I first transposed the images and values in the nested lists, to get only the list of images Using list comprehension. 
To create the collage I called the collage function on the RGB image lists. I thought it was best to use the actual width of the images we already have for the ‘frame width’ argument since we are resizing images. The third argument was the image per row which I set to 3.
Conclusions and Reflections
Image manipulation is a method of data process, just like any other data, the error is shown when running the code explains what should change. when attempting to create a collage from the flat lists I had a ‘TypeError: unhashable type: 'PngImageFile'’. I figured that the problem is the file type, so I initiated a new list and appended to it the images from the flat list after converting them to RGB images using PIL.









REFERENCES
Montfort, N. (2016). Exploratory programming for the arts and humanities.