Text processing
Introduction
Data Visualization is an approach to exploring. Visualizing simplifies abstract data, shows patterns and give the option of observing data from different angles. “visualization, a designed way of representing data to allow a viewer to better explore and understand it.” (Montfor, 2016, p. 223).
Approach
For the first 2 requirements, I thought I’d have 2 python files, one to define all functions I need to use for both requirements on and the other file is to extract data from both book chapters and call my pre-defined functions to return the required results and print them out to 2 csv files, one for words and one for sentences. The csv files would be later on used to have a pandas data-frame that I would use to visualize the data into an interactive table and a bar chart
Methodology approach:
The functions file:
1. extracting the longest sentences from both chapter
I started with defining function, the first function was the word counter that will count the number of words.
For extracting the longest sentences, I thought the best is to define a function that would return a dictionary that has the 10 longest sentences as keys, and chapter numbers as values. Within the function I used regular expression to compile the sentence ender like a dot or a question mark, to decide where to split the text to have a list of sentences and accordingly write them into a dictionary. Then I ordered the ‘keys’ of the dictionary from longest to shortest using a word
count lambda function. Accordingly, I used a for loop that would have only the 10 longest sentences in the chapter.
2. Extractthetoptenmostfrequentwordsandtheirlength.
I started with defining a function that would split my text into words using regular expressions to specify the criteria for splitting the chapter into words. Then filter the stop words to remove them and return a list of words without the stop- words, taking into consideration case sensitivity.
I then used the counter function and the most common method to get the frequency of the words in the chapter. Then I converted the data into the dictionary with the most frequent words as keys chapter numbers as values.
The task file:
1. extracting the longest sentences from both chapter:
I started with opening both the book chapter sand the stop words list files using the open () function and the read () method to read their content into 2 variables. I also used the open function to read out the content of the stop words into a variable.
I started with calling the first function of splitting text into sentences and created 2 dictionaries that each has to longest 10 sentences from each chapter.
I then updated the 1st dictionary content with the second one to have all the content in one dictionary, since I specified the value of each key to be the chapter number. Then I ordered both of them using the sorted method and a word count lambda function.
I then opened a new CSV file to write the extracted data to use the CSV writer function. I specified the data I need to add to the CSV file using a for loop and then used the ‘writerow’ function to write rows in the CSV file to have 4 columns, ranking, sentence, length and chapter.
2. Extract the top ten most frequent words and their length.
For this requirement, I called the pre-defined functions from the functions file, starting with the function that removes stop-words and returns a dictionary of words as keys and chapter numbers as values. I then merged the content of both dictionaries just like I did for the sentences. And similarly, created a CSV file that contains 4 columns word, frequency, length, and chapter. I called the frequency

function to get the frequency of each ‘key’ or word from the dictionary in each chapter.
3. GeneratevisualizationusingDash
For this requirement, I created 2 files, one for words and one for sentences.
I started with importing data from the CSV files into a pandas data frame. I had a problem with filtering content so for each time I need to view data from a chapter, I need to specify it and refresh the interface to reflect the new result for the words file, but I used a different approach for the sentences file. I created an app, an interactive user interface, where users can filter the data table and selected what data to visualize in the bar chart. For the layout, I used dash and HTML to modify the styling of the page. For the sentences, the bar chart displays the longest sentences on the ‘x’ axis and the length on the ‘y’ axis. As for the words’ visualization, the ‘x’ axis contains the words and the ‘y’ axis contains the frequency of the words.
Conclusions and Reflections
The visualization could be done differently, there are several methods to visualize data according to the datatype.
Reference:
Montfort, N. (2016). Exploratory programming for the arts and humanities.